# api/openapi.yaml
openapi: 3.1.1
info:
  title: washmonitor-agent-api
  description: This is the API served by the washmonitor agent
  version: 1.0.0
servers:
  - url: http://localhost:8001
paths:
  /washer/setAgentStatus:
    post:
      summary: Set the status of the washer agent to either 'monitor' or 'idle'
      description: This endpoint allows the user to set the status of the washer agent to either 'monitor' or 'idle'.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [status, user]
                  properties:
                    status:
                      type: string
                      enum: [monitor]
                      description: The status must be 'monitor'
                    user:
                      type: string
                      description: Numeric user ID as a string (for example, '1')
                      pattern: "^[1-6]$"
                - type: object
                  required: [status]
                  properties:
                    status:
                      type: string
                      enum: [idle]
                      description: The status must be 'idle'
                    user:
                      type: string
                      description: The user (optional if status is 'idle')
      responses:
        "200":
          description: The status of the washer agent has been set
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Agent status set successfully
                  status:
                    type: string
                    example: monitor
                  user:
                    type: string
                    example: "1"
        "400":
          description: The request was malformed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    examples:
                      - Malformed request
                      - User is required when status is 'monitor'
                      - User must be a numeric ID
                      - User ID is not configured
        "500":
          description: An error occurred while setting the agent status
  /washer/getAgentStatus:
    get:
      summary: Get the status of the washer agent
      description: This endpoint allows the user to get the status of the washer agent
      responses:
        "200":
          description: The status of the washer agent has been retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: idle
                  user:
                    type: string
                    example: ""
        "500":
          description: An error occurred while getting the agent status
  /washer/checkin:
    post:
      summary: Agent heartbeat for washer
      description: Used by the washer agent to announce it's alive. Updates last-seen timestamp and sensor status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sensorOk]
              properties:
                sensorOk:
                  type: boolean
                  description: Whether the washer camera feed is reachable
              example:
                sensorOk: true
      responses:
        "200":
          description: Check-in received
        "400":
          description: Malformed request, missing sensorOk
        "500":
          description: Server error
  /dryer/setAgentStatus:
    post:
      summary: Set the status of the dryer agent to either 'monitor' or 'idle'
      description: This endpoint allows the user to set the status of the dryer agent to either 'monitor' or 'idle'.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [status, user]
                  properties:
                    status:
                      type: string
                      enum: [monitor]
                      description: The status must be 'monitor'
                    user:
                      type: string
                      description: Numeric user ID as a string (for example, '1')
                      pattern: "^[1-6]$"
                - type: object
                  required: [status]
                  properties:
                    status:
                      type: string
                      enum: [idle]
                      description: The status must be 'idle'
                    user:
                      type: string
                      description: The user (optional if status is 'idle')
      responses:
        "200":
          description: The status of the dryer agent has been set
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Agent status set successfully
                  status:
                    type: string
                    example: monitor
                  user:
                    type: string
                    example: "1"
        "400":
          description: The request was malformed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    examples:
                      - Malformed request
                      - User is required when status is 'monitor'
                      - User must be a numeric ID
                      - User ID is not configured
        "500":
          description: An error occurred while setting the agent status
  /dryer/getAgentStatus:
    get:
      summary: Get the status of the dryer agent
      description: This endpoint allows the user to get the status of the dryer agent
      responses:
        "200":
          description: The status of the dryer agent has been retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: idle
                  user:
                    type: string
                    example: ""
        "500":
          description: An error occurred while getting the agent status
  /dryer/checkin:
    post:
      summary: Agent heartbeat for dryer
      description: Used by the dryer agent to announce it's alive. Updates last-seen timestamp and sensor status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sensorOk]
              properties:
                sensorOk:
                  type: boolean
                  description: Whether the vibration sensor has submitted state recently
              example:
                sensorOk: true
      responses:
        "200":
          description: Check-in received
        "400":
          description: Malformed request, missing sensorOk
        "500":
          description: Server error
  /users/names:
    get:
      summary: Get the list of user names and colors
      description: This endpoint returns configured users keyed by numeric user ID strings.
      responses:
        "200":
          description: The list of user names and colors has been retrieved
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: "#/components/schemas/UserInfo"
              examples:
                dynamicUsers:
                  summary: Dynamic users keyed by ID
                  value:
                    "1":
                      name: John
                      color: "#3b82f6"
                    "2":
                      name: Jane
                      color: "#22c55e"
  /health:
    get:
      summary: Health check including agent availability
      description: Returns server health and agent last-seen/online status.
      responses:
        "200":
          description: Health information
          content:
            application/json:
              schema:
                type: object
                properties:
                  api:
                    type: object
                    properties:
                      status:
                        type: string
                        example: ok
                  washer:
                    type: object
                    properties:
                      online:
                        type: boolean
                        example: true
                      lastSeen:
                        type: string
                        format: date-time
                        example: "2026-03-17T12:34:56Z"
                      sensor:
                        type: object
                        properties:
                          online:
                            type: boolean
                            description: Whether the sensor is reporting within the offline threshold
                            example: true
                          ok:
                            type: boolean
                            description: Last reported sensor health from the agent
                            example: true
                          lastSeen:
                            type: string
                            format: date-time
                            example: "2026-03-17T12:34:56Z"
                  dryer:
                    type: object
                    properties:
                      online:
                        type: boolean
                        example: true
                      lastSeen:
                        type: string
                        format: date-time
                        example: "2026-03-17T12:34:00Z"
                      sensor:
                        type: object
                        properties:
                          online:
                            type: boolean
                            description: Whether the sensor is reporting within the offline threshold
                            example: true
                          ok:
                            type: boolean
                            description: Last reported sensor health from the agent
                            example: true
                          lastSeen:
                            type: string
                            format: date-time
                            example: "2026-03-17T12:34:00Z"

components:
  schemas:
    UserInfo:
      type: object
      properties:
        name:
          type: string
          description: The display name of the user.
          example: Jane
        color:
          type: string
          description: Color hex code
          example: "#22c55e"